requirements.txt
asgiref==3.9.1
Django==5.2.5
django-qr-code==4.2.0
Markdown==3.8.2
packaging==25.0
pillow==11.3.0
psycopg2==2.9.10
sqlparse==0.5.3
gunicorn
whitenoise
python-decouple
dj-database-url
